Robustcheckout updates (#89)
* Bug 1381768 - Remove robustcheckout with ssl monkeypatch

This has been unused since 95846d9e8c.

* Bug 1381768 - Vendor latest robustcheckout

From revision 0451f128550d8c9dcf83d6628dea5a8030d5cc08 of
the version-control-tools repository. This version is deployed
in mozilla-central and seems to be working fine.

* Bug 1381768 - Don't seed instance with Firefox clone

There is no advantage to doing this as part of instance
creation. The work is going to get done anyway, so we might
as well do it as part of the task so we have visibility into it.

nodeploy