NoBug - Remove git and Intel driver install; nodeploy
Git install as is hangs and prevents the manifest from being fully applied. Will fix and add back in if a need for git on Win 7 hardware testers becomes apparent. 

At this time the Intel driver needs to be present on the machine prior to OCC being ran. This is because human interaction is needed to get the VMs to use the Intel driver. This will be reworked and automated in the future. At which time it will be reintroduced to OCC. 

Committing with out review because this manifest is not in production use at this time.