Adding noautoupdate reg entry
This is 2 fold to prevent updates on WIn 10 nodes. pauseupdate will pause updates, and Donnotupdate will prevent the updates from being downloaded with out human interaction if the pause is turned off. This tricky because the Windows Automatic Update service needs to run for OCC, so at the time of the run the machine is vulnerable to updates.  Updates cause 2 immediate problems; They cause an increase in disk space and can knock the Intel card out of use.