Bug 1451837 - run dism command to clean up
After disk space less that 25 GB run a dism command. This command will clean up C:\Windows\WinSxS among other items. I putting this outside of the cleanup portion, so it will not run on every time generic worker exits. When this was the case we saw some odd behavior with this command.