bug 1529919 - remove unused secrets. (#224)
* bug 1529919 - remove unused secrets.

We switched over to the secrets service for gapi.token and
mozilla-{desktop,fennec}-geoloc.api.key. It looks like `mozilla-api.key`
is a copy of the fennec key, and is unused.

I'm fairly certain we're also not using the rest of the files removed
here -- searchfox isn't showing me usage for any of these. We may want
to be a bit careful with this rollout, however.

* readd tc-sccache.boto